feat- 초기 구조 설정
