+++ build.gradle
@@ -33,6 +33,7 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
 
+
 	// 데이터베이스 쿼리 로깅 용 라이브러리
 	implementation 'p6spy:p6spy:3.9.1'
 }
