--- src/main/java/spharos/msg/domain/bundle/Bundle.java
@@ -1,4 +1,0 @@
-package spharos.msg.domain.bundle;
-
-public class Bundle {
-}
