+++ src/main/java/spharos/msg/domain/bundle/controller/BundleController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.bundle.controller;
+
+public class BundleController {
+}
