+++ src/main/java/spharos/msg/domain/bundle/entity/Bundle.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.bundle.entity;
+
+public class Bundle {
+}
