+++ src/main/java/spharos/msg/domain/cart/repository/CartRepository.java
@@ -1,7 +1,7 @@
 package spharos.msg.domain.cart.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import spharos.msg.domain.cart.Cart;
+import spharos.msg.domain.cart.entity.Cart;
 
 public interface CartRepository extends JpaRepository<Cart,Long> {
 }
