+++ src/main/java/spharos/msg/domain/cart/service/CartService.java
@@ -1,13 +1,18 @@
 package spharos.msg.domain.cart.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import spharos.msg.domain.cart.dto.CartRequestDto;
+import spharos.msg.domain.cart.repository.CartRepository;
 
 @Service
+@RequiredArgsConstructor
 public class CartService {
-    public void createCart(CartRequestDto cartRequestDto) {
-
+    private final CartRepository cartRepository;
+    public void addCart(CartRequestDto cartRequestDto) {
+        //상품 옵션 없는 경우 바로 add
+        //
     }
 
     public void getCart(UserDetails userDetails) {
