--- src/main/java/spharos/msg/domain/category/Category.java
@@ -1,4 +1,0 @@
-package spharos.msg.domain.category;
-
-public class Category {
-}
