+++ src/main/java/spharos/msg/domain/category/controller/CategoryController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.category.controller;
+
+public class CategoryController {
+}
