+++ src/main/java/spharos/msg/domain/category/entity/Category.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.category.entity;
+
+public class Category {
+}
