--- src/main/java/spharos/msg/domain/coupon/Coupon.java
@@ -1,4 +1,0 @@
-package spharos.msg.domain.coupon;
-
-public class Coupon {
-}
