+++ src/main/java/spharos/msg/domain/coupon/controller/CouponController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.coupon.controller;
+
+public class CouponController {
+}
