+++ src/main/java/spharos/msg/domain/coupon/entity/Coupon.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.coupon.entity;
+
+public class Coupon {
+}
