--- src/main/java/spharos/msg/domain/like/Like.java
@@ -1,4 +1,0 @@
-package spharos.msg.domain.like;
-
-public class Like {
-}
