+++ src/main/java/spharos/msg/domain/like/controller/LikeController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.like.controller;
+
+public class LikeController {
+}
