+++ src/main/java/spharos/msg/domain/like/entity/Like.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.like.entity;
+
+public class Like {
+}
