--- src/main/java/spharos/msg/domain/order/Order.java
@@ -1,4 +1,0 @@
-package spharos.msg.domain.order;
-
-public class Order {
-}
