+++ src/main/java/spharos/msg/domain/order/controller/OrderController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.order.controller;
+
+public class OrderController {
+}
