+++ src/main/java/spharos/msg/domain/order/entity/Order.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.order.entity;
+
+public class Order {
+}
