+++ src/main/java/spharos/msg/domain/product/controller/ProductController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.product.controller;
+
+public class ProductController {
+}
