--- src/main/java/spharos/msg/domain/review/Review.java
@@ -1,4 +1,0 @@
-package spharos.msg.domain.review;
-
-public class Review {
-}
