+++ src/main/java/spharos/msg/domain/review/controller/ReviewController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.review.controller;
+
+public class ReviewController {
+}
