+++ src/main/java/spharos/msg/domain/review/entity/Review.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.review.entity;
+
+public class Review {
+}
