--- src/main/java/spharos/msg/domain/search/Search.java
@@ -1,4 +1,0 @@
-package spharos.msg.domain.search;
-
-public class Search {
-}
