+++ src/main/java/spharos/msg/domain/search/controller/SearchController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.search.controller;
+
+public class SearchController {
+}
