+++ src/main/java/spharos/msg/domain/search/entity/Search.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.search.entity;
+
+public class Search {
+}
