+++ src/main/java/spharos/msg/domain/users/controller/UsersController.java
@@ -1,0 +1,4 @@
+package spharos.msg.domain.users.controller;
+
+public class UsersController {
+}
