+++ src/main/java/spharos/msg/global/configuration/WebConfig.java
@@ -1,0 +1,4 @@
+package spharos.msg.global.configuration;
+
+public class WebConfig {
+}
